Revert "Feature/add marine biodiversity layer"
