+++ .nvmrc
@@ -1,1 +1,1 @@
-12.21.0
+lts/erbium
