--- _config.yml
@@ -1,1 +1,0 @@
-remote_theme: pmarsceill/just-the-docs
