+++ src/components/arcgis-layer-manager/arcgis-layer-manager.js
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { layersConfig } from 'constants/layers-config';
+import { layersConfig } from 'constants/mol-layers-configs';
 import { setLayersOrder, setLayersVisibility, updateSceneLayersBasedOnUserConfig } from 'utils/arcgis-layer-manager-utils';
 import { addActiveLayersToScene } from 'utils/layer-manager-utils';
