+++ src/constants/layers-slugs.js
@@ -80,7 +80,6 @@ export const AMPHIB_RICHNESS = 'amphib-rich';
 export const AMPHIB_PRIORITY = 'amphib-priority';
 export const FISHES_RARITY = 'fishes-rarity';
 export const FISHES_RICHNESS = 'fishes-rich';
-export const FISHES_PRIORITY = 'fishes-priority';
 export const MAMMALS_PRIORITY = 'mammals-priority';
 export const MAMMALS_RARITY = 'mammals-rare';
 export const MAMMALS_RICHNESS = 'mammals-rich';
@@ -94,12 +93,6 @@ export const CONIFERS_RICHNESS = 'conifers-rich';
 export const REPTILES_PRIORITY = 'reptiles-priority';
 export const REPTILES_RARITY = 'reptiles-rarity';
 export const REPTILES_RICHNESS = 'reptiles-richness';
-export const MARINE_MAMMALS_RICHNESS = 'marine-mammals-richness';
-export const MARINE_MAMMALS_RARITY = 'marine-mammals-rarity';
-export const MARINE_MAMMALS_PRIORITY = 'marine-mammals-priority';
-export const ALL_MARINE_VERTEBRATES_RICHNESS = 'all-marine-vertebrates-richness';
-export const ALL_MARINE_VERTEBRATES_RARITY = 'all-marine-vertebrates-rarity';
-export const ALL_MARINE_VERTEBRATES_PRIORITY = 'all-marine-vertebrates-priority';
 // Hummingbirds
 export const HUMMINGBIRDS_RARITY = 'hummingbirds-rare';
 export const HUMMINGBIRDS_RICHNESS = 'hummingbirds-rich';
