+++ src/services/metadata-service.js
@@ -1,15 +1,11 @@
 import { loadModules } from 'esri-loader';
-import {
-  METADATA_SERVICE_URL,
-  DEV_METADATA_SERVICE_URL
-} from 'constants/layers-urls';
-const SERVICE_URL = process.env.NODE_ENV === "development" ? DEV_METADATA_SERVICE_URL : METADATA_SERVICE_URL;
+import { METADATA_SERVICE_URL } from 'constants/layers-urls';
 
 function getMetadata(slug) {
   return new Promise((resolve, reject) => {
     loadModules(["esri/tasks/QueryTask", "esri/tasks/support/Query"]).then(([QueryTask, Query]) => {
       var queryTask = new QueryTask({
-        url: SERVICE_URL
+        url: METADATA_SERVICE_URL
       });
       var query = new Query();
       query.outFields = ["*"];
